Chore(#24): Auth 관련 타입 및 스키마 파일 구조 단순화
