+++ src/types/auth/index.ts
@@ -1,5 +1,2 @@
-export * from './signup.schema'
-export * from './signup.type'
-
-export * from './signin.schema'
-export * from './signin.type'
+export * from './schema'
+export * from './type'
