+++ src/types/auth/schema.ts
@@ -93,4 +93,13 @@ export const signupSchema = z
     }
   })
 
+export const signinSchema = z.object({
+  email: z
+    .string()
+    .email('올바른 이메일 형식이 아닙니다')
+    .nonempty('이메일을 입력해주세요'),
+  password: z.string().min(1, '비밀번호를 입력해주세요'),
+})
+
 export type SignupFormValues = z.infer<typeof signupSchema>
+export type SigninFormValues = z.infer<typeof signinSchema>
