--- src/types/auth/signin.type.ts
@@ -1,3 +1,0 @@
-type OAuth = 'google' | 'kakao' | 'naver'
-
-export type { OAuth }
